refactor(sanity): remove provider + context, move logic to hook
